Extract isInterceptionRouteRewrite to its own module
Move isInterceptionRouteRewrite from generate-interception-routes-rewrites.ts
to lib/is-interception-route-rewrite.ts to allow importing it without pulling
in the full interception routes generation code. Use lazy require in
manifest-loader.ts to defer loading.
